% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albums}
\alias{albums}
\title{Top 3000 Albums from tsort.info}
\format{A data frame with 3000 rows and 9 variables:
\describe{
    \item{position}{The item's position in the overall list, from 1 to 3000}
    \item{artist}{The name of the artist / band}
    \item{name}{Name of the Album}
    \item{year}{The year allocated to the entry}
    \item{final_score}{Total score, calculated using weights (factors) by region and year, as documented in more detail at tsort.info. See FAQ. Datasets containing weights are available there.}
    \item{raw_usa}{Raw score (i. e. without applying weights) from the USA based charts}
    \item{raw_eng}{Raw score (i. e. without applying weights) from the non-USA English speaking countries (UK, Canada, Australia, New Zealand, South Africa, Ireland and Hong Kong)}
    \item{raw_eur}{Raw score (i. e. without applying weights) from the non-English speaking countries in Europe (Germany, France, Austria, Norway, Sweden, Italy, Switzerland, Spain, Netherlands, Belgium, Poland, Denmark, Finland and the Vatican)}
    \item{raw_row}{Raw score (i. e. without applying weights) from the remaining countries (row = "rest of world"), i. e. exluding the markets above.}
    }}
\source{
\url{https://tsort.info}
}
\usage{
albums
}
\description{
A dataset containing the Top 3000 Albums from \url{https://tsort.info}. Version 2-8-0025.
}
\details{
Note that usage of data from tsort.info is free under the following conditions:

1. Acknowledge the source.

2. Prominently link to \url{https://tsort.info}

3. Always include version number.

File version "2-8-0025" included as an attribute. See Examples section.

Note that rankings and scores are likely to change in new versions provided at \url{https://tsort.info}.
}
\examples{
data(albums)        # Lazy loading. Data becomes visible as soon as used, e. g. via str(songs).

## Access file version number:
attr(albums, "version")
str(albums)
}
\keyword{datasets}
