% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chart2000_data.R
\name{ch2k_get_version}
\alias{ch2k_get_version}
\title{Get current file version number}
\usage{
ch2k_get_version(url = "https://chart2000.com/about.htm")
}
\arguments{
\item{url}{Website containing information on version numbers.}
}
\description{
\code{ch2k_get_version} Extracts the current file version used at \url{https://chart2000.com}.
}
\details{
Note that usage of data from chart2000.com is free under the following conditions:

1. Acknowledge the source.

2. Prominently link to \url{https://chart2000.com}

3. Always include version number.
}
\examples{
## Get current file version.
tsm_get_version()
}
\seealso{
Other read functions: 
\code{\link{ch2k_read_albums}()},
\code{\link{ch2k_read_songs}()},
\code{\link{tsm_get_version}()},
\code{\link{tsm_read_albums}()},
\code{\link{tsm_read_chart}()},
\code{\link{tsm_read_songs}()}
}
\concept{read functions}
