% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chart2000_data.R
\name{ch2k_read_albums}
\alias{ch2k_read_albums}
\title{Importing Album Data from chart2000.com}
\usage{
ch2k_read_albums(
  version = ch2k_get_version(),
  na = "unknown",
  ch2k_path = "https://chart2000.com/data/",
  ...
)
}
\arguments{
\item{version}{The version in string format. Defaults to function reading the current version from chart2000.com. Example: "0-3-0073".}

\item{na}{How unknown years are encoded, in string format.}

\item{ch2k_path}{URL to Website.}

\item{...}{Pass additional parameters to readr::read_csv. See ?read_csv for details.}
}
\description{
\code{ch2k_read_albums} reads the top 50 albums for every month starting in January 2000 published at \url{https://chart2000.com}.
A wrapper for readr::read_csv with convenient defaults, see parameters.
}
\details{
Note that usage of data from chart2000.com is free under the following conditions:

1. Acknowledge the source.

2. Prominently link to \url{https://chart2000.com}

3. Always include version number.
}
\examples{
\dontrun{
## Read Top 50 albums for each month
albums2000 <- ch2k_read_albums()

## Access file version number:
attr(albums2000, "version")
}
}
\seealso{
Other read functions: 
\code{\link{ch2k_get_version}()},
\code{\link{ch2k_read_songs}()},
\code{\link{tsm_get_version}()},
\code{\link{tsm_read_albums}()},
\code{\link{tsm_read_chart}()},
\code{\link{tsm_read_songs}()}
}
\concept{read functions}
