% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{tsm_get_version}
\alias{tsm_get_version}
\title{Get current file version number}
\usage{
tsm_get_version(url = "https://tsort.info/music/faq_version_numbers.htm")
}
\arguments{
\item{url}{Website containing information on version numbers.}
}
\description{
\code{tsm_get_version} Extracts the current file version used at \url{https://tsort.info}.
}
\details{
Note that usage of data from tsort.info is free under the following conditions:

1. Acknowledge the source.

2. Prominently link to \url{https://tsort.info}

3. Always include version number.
}
\examples{
## Get current file version.
tsm_get_version()
}
\seealso{
Other read functions: 
\code{\link{tsm_read_albums}()},
\code{\link{tsm_read_chart}()},
\code{\link{tsm_read_songs}()}
}
\concept{read functions}
