% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_ggbox}
\alias{tsm_ggbox}
\title{Plotting functions for tsort.info data}
\usage{
tsm_ggbox(
  data,
  title = "Album release years",
  subtitle = "Boxplots and raw data points",
  farben = tmaptools::get_brewer_pal("Dark2", n = 16)
)
}
\arguments{
\item{data}{Data frame to use for plotting.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{farben}{Vector of colors.}
}
\value{
Data is returned silently, so that plotting can be used within a pipe.
}
\description{
Plotting functions for tsort.info data
}
\examples{

tsm_ggbox(albums)
tsm_ggbox(albums, title = "Release years of albums by selected bands"
                  subtitle = "")
}
