% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_ggbox_plotly}
\alias{tsm_ggbox_plotly}
\title{Interactive boxplot using plotly}
\usage{
tsm_ggbox_plotly(p, tooltip = c("x", "y", "label"))
}
\arguments{
\item{p}{A ggplot2 boxplot, preferably created using tsm_ggbox()}

\item{tooltip}{Which information shall be displayed on mouseover? Choices are "x", "y", and "label"}
}
\value{
An interactive plotly boxplot
}
\description{
Interactive boxplot using plotly
}
\examples{
p <- tsm_ggbox(albums)
tsm_ggbox_plotly(p)
}
