% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{tsm_read_chart}
\alias{tsm_read_chart}
\title{Importing Full Data from tsort.info}
\usage{
tsm_read_chart(version = "2-8-0023", na = "unknown",
  tsort_path = "https://tsort.info/")
}
\arguments{
\item{version}{The version in string format.}

\item{na}{How unknown years are encoded, in string format.}

\item{tsort_path}{URL to Website.}
}
\description{
\code{read_chart} reads the complete set of data published at https://tsort.info.
}
\seealso{
Other read functions: \code{\link{tsm_read_songs}}
}
\concept{read functions}
