% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_theme}
\alias{tsm_theme}
\title{A ggplot2 theme for plotting tsort.info data.}
\usage{
tsm_theme()
}
\value{
A complete ggplot2 theme based on theme_bw(). Font size 14, text family "serif"
}
\description{
Function to be used within ggplot2 calls or ggplot2 functions within the tsortmusicr package.
}
\examples{
library(ggplot2)
p <- ggplot(albums, aes(x = year, y = final_score)) +
   geom_jitter(alpha = 0.7) +
   labs(title = "Final score by year")
p
p + tsm_theme()
}
