% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsortmusicr.R
\docType{package}
\name{tsortmusicr}
\alias{tsortmusicr}
\alias{tsortmusicr-package}
\title{tsortmusicr: Functions to work with data from http://tsort.info/ - The World's Music Charts}
\description{
Please note: Package is in an early development stage. Not ready for usage!
}
\section{Idea}{

Import data from https://tsort.info/; provide some functions for data wrangling and plotting
}

