% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_ggbox}
\alias{tsm_ggbox}
\title{Boxplots for tsort.info data}
\usage{
tsm_ggbox(
  data,
  xvar = artist,
  yvar = year,
  max_cats = 15,
  title = "Album release years",
  subtitle = "Boxplots and raw data points",
  farben = tmaptools::get_brewer_pal("Dark2", n = max_cats, plot = FALSE)
)
}
\arguments{
\item{data}{Data frame to use for plotting}

\item{xvar}{Categorical variable}

\item{yvar}{Continuous variable}

\item{max_cats}{Maximum number of categories to be displayed}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{farben}{Vector of colors. Defaults to palette "Dark2".}
}
\value{
ggplot object, so that it can be further manipulated (e. g. apply different theme, theme options)
}
\description{
Create boxplot including original data points.
}
\examples{

tsm_ggbox(albums)
tsm_ggbox(albums, title = "Release years of albums by selected bands"
                  subtitle = "")
albums \%>\%
   filter(artist != "Original Soundtrack") \%>\%
   tsm_ggbox()
}
