% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_ggbox}
\alias{tsm_ggbox}
\title{Boxplots for tsort.info data}
\usage{
tsm_ggbox(
  data,
  xvar = artist,
  yvar = year,
  label = name,
  max_cats = 15,
  title = "Album release years",
  subtitle = "Boxplots and raw data points",
  farben = tmaptools::get_brewer_pal("Dark2", n = max_cats, plot = FALSE)
)
}
\arguments{
\item{data}{data frame to use for plotting}

\item{xvar}{categorical variable}

\item{yvar}{continuous variable}

\item{label}{a variable that can be used in plotly::ggploty() to label data points}

\item{max_cats}{maximum number of categories to be displayed}

\item{title}{plot title}

\item{subtitle}{plot subtitle}

\item{farben}{vector of colors. Defaults to palette "Dark2"}
}
\value{
ggplot object, so that it can be further manipulated (e. g. apply different theme, theme options)
}
\description{
Create boxplot including original data points.
}
\examples{

library(dplyr)
tsm_ggbox(albums)
tsm_ggbox(albums, title = "Release years of albums by selected bands",
                  subtitle = "")
albums \%>\%
   filter(artist != "Original Soundtrack") \%>\%
   tsm_ggbox()
}
