% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{tsm_read_songs}
\alias{tsm_read_songs}
\title{Importing Song Data from tsort.info}
\usage{
tsm_read_songs(version = "2-8-0023", na = "unknown",
  tsort_path = "https://tsort.info/csv/")
}
\arguments{
\item{version}{The version in string format.}

\item{na}{How unknown years are encoded, in string format.}

\item{tsort_path}{URL to Website.}
}
\description{
\code{read_songs} reads the top 5000 songs.
}
\seealso{
Other read functions: \code{\link{tsm_read_chart}}
}
\concept{read functions}
