% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{tsm_read_songs}
\alias{tsm_read_songs}
\title{Importing Song Data from tsort.info}
\usage{
tsm_read_songs(
  version = tsm_get_version(),
  na = "unknown",
  tsort_path = "https://tsort.info/csv/",
  ...
)
}
\arguments{
\item{version}{The version in string format. Defaults to function reading the current version from tsort.info. Example: "2-8-0025".}

\item{na}{How unknown years are encoded, in string format.}

\item{tsort_path}{URL to Website.}

\item{...}{Pass additional parameters to readr::read_csv. See ?read_csv for details.}
}
\description{
\code{tsm_read_songs} reads the top 5000 songs.
A wrapper for readr::read_csv with convenient defaults, see parameters.
}
\details{
Note that usage of data from tsort.info is free under the following conditions:

1. Acknowledge the source.

2. Prominently link to \url{https://tsort.info}

3. Always include version number.
}
\examples{
\dontrun{
## Read Top 5000 songs
songs <- tsm_read_songs()

## Access file version number:
attr(songs, "version")
}
}
\seealso{
Other read functions: 
\code{\link{tsm_get_version}()},
\code{\link{tsm_read_albums}()},
\code{\link{tsm_read_chart}()}
}
\concept{read functions}
