% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsort_ggplot.R
\name{tsm_theme}
\alias{tsm_theme}
\title{A ggplot2 theme for plotting tsort.info data.}
\usage{
tsm_theme()
}
\value{
No return value.
}
\description{
Function to be used within ggplot2 calls or ggplot2 functions within the tsortmusicr package.
}
\examples{
tsm_ggbox() + tsm_theme()
}
