% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsortmusicr.R
\docType{package}
\name{tsortmusicr}
\alias{tsortmusicr-package}
\alias{tsortmusicr}
\title{Functions to work with data from https://tsort.info/ - The World's Music Charts}
\description{
Work in progress! Not a finished package! Functions and function names are likely to change in future versions.
}
\details{

}
\section{Data frames}{

Note that usage of data from tsort.info is free under the following conditions:

\enumerate{
    \item Acknowledge the source.
    \item Prominently link to https://tsort.info
    \item Always include version number.
}

Data frames that ship with the package:

\itemize{
    \item songs: Top 5000 Songs from tsort.info, 9 variables, version 2-8-0025.
    \item albums: Top 3000 Albums from tsort.info, 9 variables, version 2-8-0025.
    \item chart: Full data from tsort.info, 71141 rows, 17 variables, version 2-8-0025.
}
}

\section{Plans}{

Provide some functions for displaying these data.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fjodor/tsortmusicr}
  \item Report bugs at \url{https://github.com/fjodor/tsortmusicr/issues}
}

}
\author{
\strong{Maintainer}: Wolf Riepl \email{wolf.riepl@statistik-dresden.de}

}
